+++ components/Testimonials/Testimonials.tsx
@@ -1,6 +1,6 @@
 import Card from "@/components/Card/Card";
 import Section from "@/components/Section/Section";
-import { Variant } from "@/packages/types";
+import { Variant } from "@/types";
 import styles from "./Testimonials.module.scss";
 
 export default function Testimonials() {
